# hotelappi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "hotelappi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/webapi/bookings/book": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookingRequestModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/roomtypes/{id}": {
      "get": {
        "tags": [
          "RoomType"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RoomTypeModel"
            }
          }
        }
      }
    },
    "/roomtypes/available": {
      "get": {
        "tags": [
          "RoomType"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "endDate",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RoomTypeModel"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BookingRequestModel": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "roomTypeId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "RoomTypeModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "image": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}